MELANet_total/MELANet_S/requirements.txt
Pytorch
Torchvision
Numpy
Imageio
SciPy
Pillow